+++ integrations/vllm_plugin/vllm_tt/model_runner.py
@@ -88,7 +88,7 @@ from .attention import (
 )
 from .logger import tt_init_logger
 from .platform import TTConfig
-from .utils import shard_model
+from .vllm_utils import shard_model
 
 
 def add_kv_sharing_layers_to_kv_cache_groups(
